pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "pipecat-ai-mcp-server"
dynamic = ["version"]
description = "Pipecat MCP server for your AI agents"
license = "BSD-2-Clause"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.10"
keywords = ["pipecat", "mcp", "voice", "ai", "assistant", "speech"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Communications :: Conferencing",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
    "kokoro-onnx>=0.5.0,<1",
    "loguru>=0.7.0,<1",
    "mcp>=1.0.0",
    "pip>=25.3",
    "pipecat-ai[cartesia,deepgram,local-smart-turn-v3,rnnoise,runner,silero,webrtc,websocket]>=0.0.100",
    "pipecat-ai[mlx-whisper]>=0.0.100; sys_platform == 'darwin'",
    "pipecat-ai[whisper]>=0.0.100; sys_platform != 'darwin'",
    "python-dotenv>=1.0.0,<2",
    "requests>=2.32.5,<3",
]

[project.urls]
Homepage = "https://pipecat.ai"
Documentation = "https://docs.pipecat.ai/"
Source = "https://github.com/pipecat-ai/pipecat-mcp-server"
Issues = "https://github.com/pipecat-ai/pipecat-mcp-server/issues"
Changelog = "https://github.com/pipecat-ai/pipecat-mcp-server/blob/main/CHANGELOG.md"

[project.optional-dependencies]
daily = [ "daily-python~=0.23.0" ]
screen = [
    "pyobjc-framework-ScreenCaptureKit>=11.0; sys_platform == 'darwin'",
    "pyobjc-framework-Quartz>=11.0; sys_platform == 'darwin'",
    "pyobjc-framework-CoreMedia>=11.0; sys_platform == 'darwin'",
]

[dependency-groups]
dev = [
    "build~=1.2.2",
    "pip-tools~=7.4.1",
    "pyright>=1.1.404,<1.2",
    "ruff>=0.12.11,<1",
    "setuptools~=78.1.1",
    "setuptools_scm~=8.3.1",
]

[project.scripts]
pipecat-mcp-server = "pipecat_mcp_server.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
local_scheme = "no-local-version"
fallback_version = "0.0.0-dev"

[tool.ruff]
exclude = [".git"]
line-length = 100

[tool.ruff.lint]
select = [
    "D", # Docstring rules
    "I", # Import rules
]
ignore = [
    "D105",  # Missing docstring in magic methods (__str__, __repr__, etc.)
]

[tool.ruff.lint.per-file-ignores]
# Skip D104 (missing docstring in public package) for __init__.py files
"**/__init__.py" = ["D104"]
